make separate signal module
