+++ packages/core/index.mjs
@@ -4,6 +4,7 @@ import Fraction from './fraction.mjs';
 export {Fraction};
 export * from './hap.mjs';
 export * from './pattern.mjs';
+export * from './signal.mjs';
 export * from './state.mjs';
 export * from './timespan.mjs';
 export * from './util.mjs';
